+++ sources/render/render.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   render.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dmitry <dmitry@student.42.fr>              +#+  +:+       +#+        */
+/*   By: lorphan <lorphan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/14 22:17:43 by lorphan           #+#    #+#             */
-/*   Updated: 2022/03/31 01:01:02 by dmitry           ###   ########.fr       */
+/*   Updated: 2022/03/31 20:33:06 by lorphan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,9 +28,9 @@ void	render(t_minirt *minirt)
 			dir = create_ray(minirt, (double)x, (double)y);
 			color = raytrace(minirt, &minirt->scene->camera->pos, &dir);
 			my_mlx_pixel_put(minirt->image, x, y,
-				((color.r & 0xFF) << 16)
-				+ ((color.g & 0xFF) << 8)
-				+ (color.b & 0xFF));
+				(((int)color.r & 0xFF) << 16)
+				+ (((int)color.g & 0xFF) << 8)
+				+ ((int)color.b & 0xFF));
 			++y;
 		}
 		++x;
